make OAuth2 code independent of HTTP client
